requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
neo4j
jupyter
python-dotenv
certifi